requirements.txt
scapy>=2.5.0
mido>=1.2.10
python-rtmidi>=1.5.8
PyYAML>=6.0.1
